Finalise brute force method of intersection
This is my answer to a weekly puzzle given. I wrote suggestions for improvements. But given the data set it hardly made sense to optimise it. If the data set was expected to be mega huge, then yes, totally makes sense to optimise. This is not the shortest code around I'm sure, but it does work.